requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
colorthief==0.2.1
Pillow==10.4.0
libsql-experimental==0.0.47
resend==2.0.0
geopy==2.4.1
lxml==5.2.2
python-dotenv==1.0.1
